feat: update the panel component
