+++ projects/bulma-app/src/app/panel/panel.component.html
@@ -2,13 +2,10 @@
 <br>
 <div class="parent-container">
   <div class="component-container">
-    <bu-panel >
+    <bu-panel>
       <bu-panel-heading>
         Panel Header
       </bu-panel-heading>
-    
-      
-    
       <bu-panel-block>
         Block 1
       </bu-panel-block>
@@ -18,10 +15,7 @@
       <bu-panel-block>
         Block 3
       </bu-panel-block>
-    
     </bu-panel>
-    
-
   </div>
   <div class="code-snippet-container">
     <markdown [src]="'assets/docs/panel/panel-example.md'"></markdown>
@@ -31,13 +25,10 @@
 
 <div class="parent-container">
   <div class="component-container">
-    <bu-panel buPanel theme="primary">
+    <bu-panel theme="primary">
       <bu-panel-heading>
         Panel Header
       </bu-panel-heading>
-    
-      
-    
       <bu-panel-block>
         Block 1
       </bu-panel-block>
@@ -47,10 +38,7 @@
       <bu-panel-block>
         Block 3
       </bu-panel-block>
-    
     </bu-panel>
-    
-
   </div>
   <div class="code-snippet-container">
     <markdown [src]="'assets/docs/panel/primary-panel.md'"></markdown>
@@ -59,13 +47,10 @@
 
 <div class="parent-container">
   <div class="component-container">
-    <bu-panel buPanel theme="link">
+    <bu-panel theme="link">
       <bu-panel-heading>
         Panel Header
       </bu-panel-heading>
-    
-      
-    
       <bu-panel-block>
         Block 1
       </bu-panel-block>
@@ -75,10 +60,7 @@
       <bu-panel-block>
         Block 3
       </bu-panel-block>
-    
     </bu-panel>
-    
-
   </div>
   <div class="code-snippet-container">
     <markdown [src]="'assets/docs/panel/link-panel.md'"></markdown>
@@ -86,13 +68,10 @@
 </div>
 <div class="parent-container">
   <div class="component-container">
-    <bu-panel buPanel theme="info">
+    <bu-panel theme="info">
       <bu-panel-heading>
         Panel Header
       </bu-panel-heading>
-    
-      
-    
       <bu-panel-block>
         Block 1
       </bu-panel-block>
@@ -102,10 +81,7 @@
       <bu-panel-block>
         Block 3
       </bu-panel-block>
-    
     </bu-panel>
-    
-
   </div>
   <div class="code-snippet-container">
     <markdown [src]="'assets/docs/panel/info-panel.md'"></markdown>
@@ -113,13 +89,10 @@
 </div>
 <div class="parent-container">
   <div class="component-container">
-    <bu-panel buPanel theme="success">
+    <bu-panel theme="success">
       <bu-panel-heading>
         Panel Header
       </bu-panel-heading>
-    
-      
-    
       <bu-panel-block>
         Block 1
       </bu-panel-block>
@@ -129,10 +102,7 @@
       <bu-panel-block>
         Block 3
       </bu-panel-block>
-    
     </bu-panel>
-    
-
   </div>
   <div class="code-snippet-container">
     <markdown [src]="'assets/docs/panel/success-panel.md'"></markdown>
@@ -140,13 +110,10 @@
 </div>
 <div class="parent-container">
   <div class="component-container">
-    <bu-panel buPanel theme="warning">
+    <bu-panel theme="warning">
       <bu-panel-heading>
         Panel Header
       </bu-panel-heading>
-    
-      
-    
       <bu-panel-block>
         Block 1
       </bu-panel-block>
@@ -156,10 +123,7 @@
       <bu-panel-block>
         Block 3
       </bu-panel-block>
-    
     </bu-panel>
-    
-
   </div>
   <div class="code-snippet-container">
     <markdown [src]="'assets/docs/panel/warning-panel.md'"></markdown>
@@ -167,13 +131,10 @@
 </div>
 <div class="parent-container">
   <div class="component-container">
-    <bu-panel buPanel theme="danger">
+    <bu-panel theme="danger">
       <bu-panel-heading>
         Panel Header
       </bu-panel-heading>
-    
-      
-    
       <bu-panel-block>
         Block 1
       </bu-panel-block>
@@ -183,10 +144,7 @@
       <bu-panel-block>
         Block 3
       </bu-panel-block>
-    
     </bu-panel>
-    
-
   </div>
   <div class="code-snippet-container">
     <markdown [src]="'assets/docs/panel/danger-panel.md'"></markdown>
