--- projects/ngx-bulma/src/lib/bulma-panel/index.ts
@@ -1,5 +1,0 @@
-export * from './bulma-panel.component';
-export * from './bulma-panel.module';
-export * from './bulma-panel-block/bulma-panel-block.component';
-export * from './bulma-panel-heading/bulma-panel-heading.component';
-export * from './bulma-panel-tabs/bulma-panel-tabs.component';
