+++ projects/ngx-bulma/src/lib/panel/index.ts
@@ -1,6 +1,4 @@
 export * from './panel.component';
 export * from './panel.module';
-export * from './panel.directive';
 export * from './panel-block/panel-block.component';
 export * from './panel-heading/panel-heading.component';
-export * from './panel-tabs/panel-tabs.component';
